migration_app/requirements.txt
boto3==1.34.0
psycopg2-binary==2.9.9
azure-cosmos==4.5.1
python-dotenv==1.0.0